SNES: Set up matrix-free *after* copying the user Jacobian to snes->pc
A bug was introduced in ddd0b78 that broke nonlinear left preconditioning
due to the MATMFFD instance set up on the outer solver being copied to the
inner. The solver calls SNESSolve() on itself in a loop trying to difference
the preconditioned function. The solution is to set up the matrix-free context
after copying all relative state onto snes->pc, preserving the use case fix
that was the inspiration for the change in ddd0b78.