Make sure data on CPU is up to date.
If VecCUSPCopyFromGPUSome_Public cannot be called we have to call
VecCUSPCopyFromGPU. Otherwise the host data might be out of data or
unallocated, as is the case in the new test introduced in this branch.

With this fix the new test runex4_cusp2 passes.  There are some valgrind
errors due to uninitialized values. These are fixed by the additional
bug fixes in fix-vecscattercusp_stos and fix-cusp-dmda.