Add KSPFCG
This adds the Flexible Conjugate Gradient Method as described and analysed in Notay's 2000 paper. Most of the available options concern storage and use of previous search directions for orthogonalization. Notay's recommended truncation strategy as well as a standard truncation strategy are included.

Remaining to investigate/implement are more options for the orthogonalization procedure (currently standard Gram-Schmidt is implemented), and efficient implementation of residual monitoring (residuals norms are currently computed explicitly).