Vec: restore support for PETSC_VIEWER_NATIVE
It is useful for debugging to be able to write a DMDA Vec in PETSc
ordering instead of in native ordering.  This functionality was removed
in

  commit 742cafdd48f845c927348e65bdee73999c739268
  Author: Shri Abhyankar <abhyshr@mcs.anl.gov>
  Date:   Thu Jul 22 14:57:46 2010 -0500

      Removed PETSC_VIEWER_NATIVE related code

and discussed on petsc-dev in

  http://mid.mail-archive.com/8738qk1d3u.fsf@mcs.anl.gov

This commit partially reverts the above and adds a test.
PETSC_VIEWER_NATIVE is the documented way to write the Vec contents
without converting to natural ordering.  This interface should be
preserved until there is a cleaner alternative.

Reported-by: Gautam Bisht <gbisht@lbl.gov>