Nightlysummary: Incomplete builds are now included in the summary.
Incomplete builds do not necessarily include all .log-files, for which
reason they might have not shown up in the summary table. With this
commit the presence of build_*.log is taken as the main indicator
for the availability of a test configuration. If other log files
are missing, the respective table row contains 'Incomplete'.
Based on suggestions by Satish.