make: PETSC_ARCH/conf/files depends on PETSC_ARCH/include/petscconf.h
When a PETSC_ARCH is reconfigured with an additional external package,
petscconf.h is updated and we expect conf/files to be regenerated, but
there was no explicit dependency causing that to happen.  I think this
is the cause of the non-reproducible build failures that have been
reported.  This commit adds the explicit dependency.

Reported-by: Kai Song <ksong@lbl.gov>
Reported-by: Åsmund Ervik <asmund.ervik@ntnu.no>