DM: put constraint hooks into DMGlobalToLocalEnd() and DMLocalToGlobalBegin().
I put this directly into the functions instead of into the linked list
of hooks.  My reasoning is that if constraints are present, they should
be considered atomic with the PetscSFBcast()/PetscSFReduce(), and that
allowing other user-defined hooks to come between the communication and
the constraint hooks could result in the user trying to access or set
constrained values before they are defined.