Plex: rename Constraints as Anchors for clarity.
Now that there are DMGet/SetConstraints() functions, some clarity is
needed:

DMGet/SetConstraints() are any type of linear constraints that are
considered part of the global-to-local map.

DMPlexGet/SetAnchors() are linear constraints that make a point's dofs
linear dependent on the dofs of a set of anchor points.

DMPlexGet/SetTree() are anchor constraints that derive from a
hierarchically refined mesh.