DM+Plex: change the collective status of default constraints, anchors, and reference tree
The default constraints, anchors, and reference tree are changed to have
PETSC_COMM_SELF communicators.

The reasoning is that we would like tree constraints to cause only
minimal changes to the solution workflow demonstrated in
snes/examples/tutorials/ex12.c.  We would like to insert only the
following code:

  /* ... (construct/choose a reference tree) */
  DMPlexSetReferenceTree(dm,refTree);
  /* ... (construct the tree structure) */
  DMPlexSetTree(dm,...);

This means that the presence of some type of constraints may not be
detected until DMPlexMatSetClosure(), within a cell loop.  If that
happens, the following call graph occurs:

DMPlexMatSetClosure(dm,...) calls
  DMGetDefaultConstraints(dm,...) calls
   *DMCreateDefaultConstraints_Plex(dm,...) calls
      DMPlexGetAnchors(dm,...) calls
       *DMPlexCreateAnchors_Tree(dm,...)
      DMPlexComputeAnchorMatrix_Tree_FromReference(dm,...) calls
        DMGetDefaultConstraints(refTree,...) calls
         *DMCreateDefaultConstraints_Plex(refTree,...) calls
            DMPlexGetAnchors(refTree,...) calls
             *DMPlexCreateAnchors_Tree(refTree,...)

The objects created in this cascade have to be local.