Plex: added DMPlexGet/SetMaxProjectionHeight()
The idea is to accomodate meshes (such as non-conforming meshes), where
a global basis function restricted to a cell may not correspond to a
single basis function on the cell, but to a linear combination.  The
most general case would be to do an L2 projection, but we would have to
find a way to invert control to use a KSPSolve().  This approach at
least expands the available set of local projection operations.