PetscDualSpace: added PetscDualSpaceGetHeightSubspace()
We already acknowledge that a functional in the dual space may have its
quadrature points limited to a mesh point in the boundary of a cell.
This goes one step further and reproduces those functionals in terms of
a reference element for the lower-dimensional mesh points.  This was
written to be used for function projection for non-conformal meshes:
a global dof may not align with any local cell dof, but may align with a
facet or edge dof.  Something like this would also be necessary for
implementing Scott-Zhang interpolation
(http://dx.doi.org/10.1090/S0025-5718-1990-1011446-7).