Plex: Build complete overlap label and stratify resulting SF.
Instead of building an SF of overlap contributions and merging it with
locally owned points, we now build a label of overlap and owned points
and stratify the resulting SF via a new utility function.

Note: This changes the point numbering of overlapping plexes, and thus
requires a test update for ex12.