Plex: Enabling parallel re-partitioning via DMPlexDistribute
This commit adds several fixes that allow DMPlexDistribute to be used on
an already distributed plex. It fixes the partition graph generation
and stratifies the migration SF used in DMPlexDistribute.

It also changes the point ordering again by always putting local points
first in DMPlexPartitionLabelInvert. This is done to deal with ParMetis
always returning the cell partitioning in local numbers.