PCASM: Allow different submatrix types
The implementation involves a call to MatConvert.

Note the new command line option, which allows things like
-pc_asm_sub_mat_type seqaijviennacl
-pc_asm_sub_mat_type aijviennacl
In the latter case, it is assumed that the sequential version of a base name
is intended.

A test is added in src/ksp/ksp/examples/tutorials, but not added to any sets of tests (in anticipation of future set of tests for viennacl-enabled configurations).