This is in response to Bitbucket issue 77 Python function readVec does not allocate memory correctly and fails for large vectors for PetscBinaryIO.py
As written PetscBinaryIO.py did not support the direct user calling of readVec, readMat, or readIS since they require the classid be read from the file before being called. I have added readObjectType() and added to the documentation so that these routines can now be called directly by the user.

Reported-by: Anonymous