Move MG and GAMG implementation headers to petsc-private
I want to be able to register GAMG plugins.  I considered a few options:

a) Move the implementation headers to petsc-private so they can be
installed

b) Make Get/Set methods for all of the GAMGOps: graph, coarsen,
optprol, etc.

c) Add a PCGAMGGet/SetOperation like MatShellGet/SetOperation for each
of these operations.

The easiest was (a), and the GAMG methods are so context-specific I'm
not sure who would benefit from their being part of the public petsc
interface.  If GAMG get's refactored to where the operations are easier
to understand on their own, then maybe (b) or (c) would be a better
option.