KSPComputeExplicitOperator: use KSP_PCApplyBAorAB
The old version always computed the left-preconditioned operator
regardless of the configuration being used.  This also did not include
diagonal scaling, if applicable.  When trying to analyze the operator
seen by the Krylov method, it is desirable to reproduce exactly the same
operator.  For most methods, that operator is applied by
KSP_PCApplyBAorAB().