Fix for ISGlobalToLocalMappingApply() for when it has a block size > 1
The issue is related with the convention used for the computation of the largest index in global ordering. In ISGlobalToLocalMappingSetUp_Private "end" is the largest index in global ordering, so the couple "start,end" defines a closed interval.

When it comes to map from global to local in ISGlobalToLocalMappingApply, indices larger the bs*end are not mapped back to local ordering, and thus the last bs-1 global indices are not mapped back.

Contributed-by: Stefano Zampini  <stefano.zampini@gmail.com>