MatStash BTS: add MAT_SUBSET_OFF_PROC_ENTRIES, impl with neighbor-only comm
Add MatOption to imprint a communication pattern on the first assembly
and reuse on subsequent assemblies.  Messages of size 0 are sent if no
values need to be communicated to that rank.  Since there are no
collectives (even Ibarrier) and no separate headers, this should be more
efficient for large process counts.  The BTS implementation must
currently be activated via the -matstash_bts option.