Add the possibility of writing HDF5 data in single precision
This commit adds the possibility for the user, at run-time,
select to write single precision data into the HDF5 file.
This is useful in many cases, as one often stop the (iterative)
solver when an approximate answer is found, thus not utilizing
full double precision. Using single precision for IO can then
save you half the disk space (or give you the possibility of
writeing twice the amount of data on the same space).

The HDF5 library takes care of the type conversion, hence the
required changes to the IO routines are very small.