Added SP output to ex47 test
Added SP output to ex47 test. The test does not check if the output
is actually single precision, it just does a roundtrip and
test if the dataset read in is the same as the dataset written
out (to a certain numerical tolerance). This test could perhaps
be made a bit more sophisticated, with more "exciting" contents
of the vector, but that is not the purpose of this commit/branch.