TSEvent: Modified TSEvent event incidence condition.
Previously, for an event to occur two conditions had to be satisfied:
(a) a change in function sign between consecutive time-steps, AND
(b) the event function value within the specified tolerance.

The new implementation only uses condition (b), i.e., an event
is detected if its function value is within the specified tolerance.
Doing so, eliminates cases where TS used to take ridiculously small
time steps (~ 1e-15) just because the function value had not changed
sign.

An additional check had to be added to the event location code
so that the same event is not detected again.