TSEvent: Fix TSEvent when an event occurs at the last time-step.
TSEvent used to not work correctly when the event was detected at the
final TS time step. This was because TS set TS_CONVERGED_TIME flag
causing the time-loop to exit inspite of the event being detected.
A reproducible example for this scenario (pointed by Hong) was
./ex40 -ts_final_time 4.1 -ts_exact_final_time MATCHSTEP. An event occured
at the last step that was not getting correctly processed by TSEventMonitor.

Fixed this issue by setting ts->reason to TS_CONVERGED_ITERATING whenever
a step is rollbacked in TSEventMonitor.