MatGetSchurComplement: revert test from LUMP back to DIAG and update output
Previous maintenance of this example inadvertently changed from DIAG to
LUMP, rendering the test output incorrect.  Subsequently, the ex21_1
output was updated to that associated with LUMP.  This commit goes back
to the original DIAG.  It also corrects the (parallel) output for the
MAT_REUSE_MATRIX test after matrix modification since the original
output was incorrect in parallel.

Reported-by: Barry Smith <bsmith@mcs.anl.gov>