BuildSystem: check for positiion-independent code after shared linker
The only sure way to know if -fPIC is working is to compile a shared
object with undefined symbols, but to find how to build a shared
library, we need to be able to compile PIC objects.  So we have to test
both at the same time (at least for one language: PETSc default C
language).  Then, once we have found a working shared linker, we can
test the other languages to see if they can compile shared objects with
undefined symbols.