Fix TSARKIMEX to work with TSEvent
arkimex was broken when using TSEvent because whenever an event was located and the step rolled back arkimex would
think that it needs to reinitialize since the solution vector has changed. Fixed this issue by adding a check for
the status of the event flag. Whenever the flag TSEVENT_PROCESSING is set, denoting the event is being processed, arkimex
will not try to reinitialize.