Commit d393f49382b06ba8f1c95499145639e321f64e19 broke -ksp/snes/ts_monitor_solution and related monitor routines since no viewer was created when the option was proccessed.
This commit fixes those bugs and provides a simplier general approach for handling viewers for standard PETSc monitors
The major change is many of the -ksp/snes/ts_monitor* functions now take the [viewertype[:filename[:viewerformat]] parameter that was previously introduced for -xxx_view options.

This commit updates about one half of the -ksp/snes/ts_monitor* options; the others that create for example a PetscDrawLG object need to be updated to take viewer objects with the correct properties set. In addition further unification/simplification of the option names must be done

Reported-by: Ed Bueler <elbueler@alaska.edu>