DMP4est: implement hierarchical transfer SFs
Each cell creates a list of the points in its closure that have dofs in
the scheme, and a PetscSFNode is created for each point in that closure.
Then, the coarse mesh is transfered to overlap the first, so that cell
overlap from fine to coarse can be determined.  Then, in looping over
the closure lists of the fine cells, we can match fine points to their
containing coarse points, which allows construction of a point-transer
SF, which can then be expanded into a dof transfer SF.

WIP DMP4est: repartitioning coarse to match fine

WIP DMP4est: factor out coarse/fine overlap code

WIP DMP4est: finally on the right track with the transfer PetscSF

DMP4est: finished DMPforetGetCellCoveringSF

WIP DMP4est: use the covering cells to compute the transfer SF

DMP4est: first completion of DMPforestGetTransferSF()