DMLabel: enable explicit control over default label values.
It seemed strange to me that DMLabelGetValue() had a default value that
means not set, but that DMLabelSetValue() could set a label to have this
value.  So now, DMLabelSetValue() will ignore attempts to set a value to
the default value, but the user can change what the default value is.

DMLabel: fix DMLabelDistribute()

The recently added default value was not transfered properly, and was
accessed by the wrong processes.

typo fix