PetscFV: added PetscFVGet/SetComponentName()
The problem right now is that fields that would be separate under
PetscFE are joined into one PetscFV field, because they are handled
together in RHS evaluation.  Maybe that can be changed in the future,
but for right now it makes sense to allow for components of FV fields to
have separate names.  I put this in the VTU viewer, but it wasn't
obvious how to do it for the VTK viewer, so I didn't do it.