Plex+SNES:always use ADD_ALL_VALUES when summing FEM element matrices
The problem is that a dof being constrained no long means just a bc dof:
it could also mean a hanging constraint. In that case that dof's
contribution can't be treated as just zero: it must be summed (via the
hanging constraint) into the residual.  The stripping of dirichlet
constraints is handled by DMLocalToGlobal().