PetscDraw X11: Eagerly destroy the popup windows in PetscDrawDestroy_X()
If the draw type ever change, the subtype should have a chance to decide
on the creation of a new popup window. If the draw type is set to null
(e.g. as a side-effect of closing the parent draw window),the X11 popup
should go away immediately (what would be the point of keeping it alive?)