PetscSFBasic: fix usage of MPI_Reduce_local
If we couldn't use one of the predefined ops, then the user must specify
an op that operates on the whole datatype passed to
PetscSFReduceBegin()/PetscSFReduceEnd().  In this setting, a block size
(link->bs) that is not equal to 1 is sometimes calculated, because it
may be determined that the datatype is in fact a contiguously packed set
of e.g., PetscInt.  That block size must be ignored in this setting,
because (link->bs * typesize) is not the data stride, typesize is.