DMP4est: added DMPforestMapCoordinates()
If we start with the coarsest possible topology, we often have a bad
geometry: e.g., the topology of a circle but the shape of a square.
p4est had a callback for mapping between the geometry defined at the
coarsest level and the desired geometry after refinement: this
promulgates that in DMP4est.