TSAdapt: Add and alternative API for weighted LTE evaluation
LTE evaluation based on TSEvaluateStep() makes sense as long as the
method contains a lower/higher-order embedded scheme. However, for some
other methods (e.g. BDF schemes), LTE evaluation is kind of more direct
(based on history rather than difference of two solutions with different
order). I'm adding an alternative API for TSAdapt to ask the TS subtype
for the weighted LTE. For now, this API is not exposed to users, however
it would make sense to do so, as TSEvaluateWLTE() would be useful for
monitoring error when running with fixed time step size.