PCML: fix MatDestroy_ML to use MatShell correctly and compatibly with visibility
MatShell interprets MATOP_DESTROY as a destructor for the context, but
the ML interface bypassed that mechanism, inverting control to call
MatDestroy_Shell statically.  However, that function is not exported by
libpetscksp.so (relevant for single-library=0), so we currently get a
link-time failure.  This commit switches to using MatShell and
MATOP_DESTROY in the usual way, so MatDestroy_Shell calls MatDestroy_ML
instead of vice-versa.