configure: improve handling of optional package dependencies [odeps]
There are 2 parts to this.
 - handle dependencies during the link test for the package
 - enable this depencency feature in the externalpackage.

Previously this was done only by GNUPackage. Now move the first part into
the common Package() class - so that CMakePackage can also use the
dependencies correctly during the link test

And update Trilinos to use odeps [aka the first feature]. The second part
is implemented manually in Trilinos.py

Reported-by: "Klinvex, Alicia Marie" <amklinv@sandia.gov>
Funded-by: IDEAS