DMP4est: use NULL geometry unless specified
Before the geometry was used in one place: vtk output.  Now it's also
used in mapping coordinates, but I want to avoid inverting the
tree-space to vertices-space map if I don't have to, i.e., when the
vertices-space map is already correct.  I now use NULL geometry to
signal this case, and create a default geometry on the fly for vtk if I
need it (it's really small, so this is fine).