SF: delay construction of rank mapping until SF type is set
To implement a shared memory optimization for PetscSF without ugly hacks
that scale poorly, we need to be aware of the shared ranks when
constructing the rank mapping.  With eager construction of rank mapping,
this means discarding that or doing fairly complicated and
memory-intense conversion.  This commit delays that construction, but
has no further behavioral change.

ex1: explicitly call PetscSFSetUp so that PetscSFView shows the full communication graph.