SFBasic: circumvent send-to-self by marking self as a distinguished group
The distinguished group is currently only self, but will later be the
group of a shared memory communicator.  The leaf buffers for
distinguished ranks are pointers directly into root buffers; MPI is not
needed.  When distinguished ranks are on different processes, we will
need a weaker synchronization mechanism to ensure memory ordering.