TSAdapt CFL: re-add guard for error message when !always_accept
The CFL implementation was never completed due to a mathematical
generality issue.  We still need to agree on a "good enough" solution
and finish the implementation, but until then, make sure it errors with
a useful message so that people don't (a) think it's expected to work or
(b) remove the error message without making it work.