PCTELESCOPE: add option to change subcommunicator type (interlaced or contiguous)
Add a new data field to PCTELESCOPE which allows, before setup only, the user to specify
which type of subcommunicator to create. Under the assumption of sequential ordering on
equally-sized compute nodes, this allows the user to specify whether their subcommunicator
lives on a subset of nodes (with the same number of ranks/node), or whether it is striped across nodes (the default).

This commit also adds some missing __FUNCT__ definitions

An example is added which calls KSPView. This isn't ideal, so a better idea of how to test this functionality would be welcome.