DMLocalizeCoordinates(): make more flexible and targeted
More flexible because methods that do calculations on facets might need
to localize them as well.  This information can be passed to the
coordinate DM with DMPlexSetMaxProjectionHeight().

More targeted by only localizing cells/points that are needed.  If I'm
localizing cells _and_ facets, then that's a lot of indiscriminate
localizing, when I only need it for a lower dimensional manifold.