DMBoundary: avoid caching the associated label
I would like to be able to add a boundary before the topology is
fixed/determined in DMForest.  This means, however, that if a boundary
is declared with a name that isn't in the existing labels, that
DMIsBoundaryPoint() can be really bad: searching for a label for every
point.  A better solution needs to be implemented.