DMP4est: construct plex data in setup
Ideally (if there are every optimized routines that don't depend of plex
data), the plex data would only be constructed as needed.  But right now
labels don't transfer until plex data is constructed, and we would like
labels to be transfered right away, so that the adaptivity forest can be
cleared right afer DMSetUp() is called on the new forest.