DMPforestGetTransferSF_Point(): try to match parents
Corner case: in AMR, the cells on either side of a facet switch from (fine,coarse) to (coarse,fine).
If we rely on the cells to map the pre-facet to the post-facet, then there's never a case where the
same cell appears in both meshes to clue us into the fact that the facets correlate.