SNES: Add flag indicating if final residual is always computed
Although SNESSolve_XXX always updates the solution variable, it may not
be the case that the residual in vec_func is the residual at the new
solution.  Track this behaviour explicitly for all concrete
implementations.  We will use this in SNESSolve_FAS to determine if the
residual must be computed before transferring it between levels.