DMLabel: switch internal storage from array to IS
This is intended to allow for more efficient storage of common label types, such as the depth label,
where each stratum is contiguous.  In the long view, I think this will be useful when making DA and
Plex use the same interface.  For a DA, it might be convenient to specify a subset (for, e.g.,
boundary conditions), using an ISStride or some kind of tbd ISSlice.