Revert Barry's commit bd373395e526a8abeca18718888f591d98e14969
It broke Jed's cool support for TS TSComputeRHSJacobianConstant
Had Jed actually added his test cases to the nightly build this
breakage would have been detected immediately but instead it had to
take years to find. Why write the test cases if you are not going to
add them to be run automatically with testing?

Update the test output for changes in formatting of various XXView commands
and added the tests to be automatically run!

Reported-by: Dominic Meiser <dmeiser@txcorp.com>