DMPlexTSComputeRHSFunctionFVM: ADD_VALUES from locF to F instead of INSERT_VALUES
We've actually designed it so that each cells contribution is only calculated once, but there are empty ghost cells on other processes with zero contributions.  There's no version of INSERT_VALUES that just inserts the local values: it's essentially non-deterministic which value gets inserted.  So we have to add all the zero contributions.