DMP4est: in setup, coordDM needs the pointSF, too.
The pointSF is inherent in the topology, which is supposed to be shared
by the DM and its coordinateDM (hence DMClone()).  But because that
topology is represented by an opaque (void *) ->data pointer (outside of
PETSc's object system), DMSetPointSF() does not actually associate the
pointSF with the shared topology.  So we have to set it multiple times.

I just talked myself out of putting a rant here, the sentiment of which
is this: DM is too big and should be refactored into smaller objects.