PetscSection: add PetscSectionSym
This object's is essentially a Petsc-idiom way of letting the user
specify implementations of

PetscSectionGetPointSyms()
PetscSectionRestorePointSyms()
PetscSectionGetFieldPointSyms()
PetscSectionRestoreFieldPointSyms()

Like PetscDualSpaceGetSymmetries(), the symmetries are descibed by R *
P, where R is a flip (or rotation with complex scalars).  I had thought
about a more managed approach, like PetscSectionGet/SetValues, which
entirely hid the existence of symmetries to begin with, but:

- then you have to differentiate between different PetscDataTypes (e.g.,
  for indices, just perform the permutation, not the flips/rotations),
  and
- what about inserting an element matrix?  When inserting an element
  matrix, it makes the most sense to apply flips/rotations to the
  element matrix, and the permutations to the indices where the matrix
  is to be inserted with MatSetValues().

so in the end I decided to just provide a simple interface to pass the
symmetries to the users and let them figure out how they want to use
them.