DMPlex: set PetscSectionSym in DMPlexCreateSectionInitial()
This just creates a PetscSectionSym from PetscDualSpaceGetSymmetries(),
with the proviso that if a section field has multiple components, we go
ahead and expand the symmetries for each component.  These expanded
symmetries are redundant, but they allow us to avoid double-loops (over
nodes and field components), when reading/writing with permutations
and/or flips.