DMPlex: use PetscSectionGet/RestorePointSyms() in DMPlexVecGet/SetClosure()
You can see that we no longer use the cone orientations directly: they
are used by the PetscSectionSym, but once the permutations/flips have
been returned, they are used.

The downside is that instead of just reading/writing backwards if o < 0,
we have indirect referencing through the permutation arrays.  But, e.g.,
for 2D, where the only permutation is equivalent to reading/writing
backward:

- there is only one permutation array, so hopefully it stays in cache
  and computing the indirection is not as bad as the worst-case for
  pointer chasing, and
- the out-or-order access is always on the element array (small,
  hopefully already in cache), not on the vector array (large, in-order
  access could improve cache performance)