DMPlex: use symmetries in DMPlexGetIndicesPoint{Fields}_Internal() and DMPlexAnchorsModifyMat()
The permutations are passed now instead of the orientations.  For
DMPlexGetIndicesPointFields_Internal(), we pass permutations for every
point (assuming PetscSectionGetFieldPointSyms() was called for multiple
points, such as for a closure set) of every field, so we have to pass an
additional parameter indicating the offset of the current point in the
set of points.