Plex tutorials ex6: use symmetries in SEM example
Other usage thus far has determined the symmetry from a dual space.  In
this example we are demonstrating SEM node ordering *without* relying on
PetscDS/PetscFE/PetscDualSpace, etc.  In this case, the user has to
describe the symmetry of their dofs for themselves.  So ex6 now
demonstrates how to describe the symmetries for tensor products grids of
degrees of freedom.