Plex-Med: Read, communicate and identify facets in parallel
The comm pattern first sends all facets to a rendezvous partition that is
defined by the original vertex partitioning. The facets are then stored
vertex-wise and communicated over the vertex SF, so that each processor
that potentially knows about a facet now has a copy. We then identify
and mark facets locally via vertex joins as before.