PetscFECreateDefault(): increase default quadrature order, make configurable
Before we ran with a default number of points per edge equal to the
polynomial order, e.g. one point for a P1 or Q1.  Gaussian quadrature
accurately integrates polynomials up to 2*K-1, where K is the number of
points per edge.  So, for P1 we can accurately integrate (v',u')
terms, but not (v',u), (v,u'), or (v,u) terms, and for Q1 we accurately
integrate none of them.  I think that having mass terms (v,u) exactly
integrated should be the default, and cheaper quadtrature should be
available on demand.

So I bumped the number of quadrature points per edge, and added the

-petscfe_default_quadatrature_order

option.