DMPlex tests ex3: update regression test for discontinuous elements
I was disconcerted that an interpolation result for p1 discontinuous
elements went from pass to fail when I changed the quadrature, but what
actually happened was that the *restriction* result went from pass to
fail.

The restriction is computed by applying the transpose of interpolation,
and then scaling the result.  The scaling of the result is designed to
be correct for a constant fine vector.  So the fine vector can be
written as f = L + C, where C is piecewise constant on each coarse cell,
and L is orthogonal to any vector that is constant on each coarse cell.
Applying the restriction operator gives Rf = RL + RC, where RC is exact.

When we underintegrated to determine the element-difference, we were
using a one-point centroid rule f().  So f(Rf - exact) = f(RL) + f(RC -
exact).  Because it is a centroid rule, f(RC-exact) = 0.  Because it is
a centroid rule and the nodes are balance around the centroid, fR is a
constant vector on each coarse cell, so f(RL) = 0.

Therefore, the fact that the restriction was exact before was an example
of superconvergence, and I'm comfortable updating the test in a way that
takes that superconvergence away.